need add cache for commands

+++ server/db/migrations/20240308082135_add_commands_table.up.sql
@@ -1,5 +1,5 @@
 CREATE TABLE "commands" (
-    "id" bigint PRIMARY KEY,
+    "id" serial PRIMARY KEY,
     "script" varchar NOT NULL,
     "description" varchar NOT NULL
 );
